test_apply_fractions is revised, passing.
but this isn't the "bug" I am looking for.  It will give a wrong value
of aadt, but I don't think I use AADT all that much, as I use VMT
instead, and that is correct---the AADT times the separate fractions of
road length sum to approximately the same thing as the AADT times the
summed fractions of road length in the grid cell.

The question i haven't sorted yet is whether AADT is used directly in
any of the summations.