more progress drilling into sql but...
no bugs.  apparent "overcount" from earlier was in fact just that I was
forgetting the year of record in my query.  account for that and I have
about the same number of records picked up my way and the simple way.

Noticed that was pullig in edge cases of roads that cross into other
counties, so I should make sure that I am limiting to the county being
analyzed.

Otherwise, still hunting for  "the bug"