more notes in the sql about Alameda.
Actually not that bad right now, but still room to improve.

HPMS is an order of magnitude lower than detector based data.

HPMS is currently less than a strict summation of HPMS, which is good.

the file TODO has some notes on what I need to do tomorrow to apply some
fixes I thought of tonight.