keep_hpms is not used, but now it is correct
not a bug, but probably confusing.  keep_hpms was empty because of how I
was computing it, so drop_hpms was always equal to detector_routes.  Now
it is still always equal to detector_routes, but at least keep_hpms is
actually an array of freeways that are not in detector_routes list.

Probably going to revisit this again and actually make drop_hpms ===
detector_routes